# coffee-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Shop API",
        "description": "API for managing coffee shop operations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "prod_name",
                                    "prod_price"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "prod_name": {
                                        "type": "string",
                                        "example": "Latte"
                                    },
                                    "prod_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "prod_description": {
                                        "type": "string",
                                        "example": "Creamy latte"
                                    },
                                    "prod_image_url": {
                                        "type": "string",
                                        "example": "https://example.com/latte.jpg"
                                    },
                                    "prod_is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "prod_meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "prod_name": {
                                        "type": "string",
                                        "example": "Updated Latte"
                                    },
                                    "prod_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "prod_description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "prod_image_url": {
                                        "type": "string",
                                        "example": "https://example.com/updated-latte.jpg"
                                    },
                                    "prod_is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "prod_meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
        "/api/tables/{id}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get table",
                "operationId": "be5358434733b508d70a9e4ea180db5d",
                "responses": {
                    "200": {
                        "description": "Table details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update table",
                "operationId": "18b8243be010c3a6b0c3b3c15ed7bb36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the table to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "table_number",
                                    "capacity"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "table_number": {
                                        "type": "string",
                                        "example": "T12"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "reserved",
                                            "out_of_service"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Table updated successfully"
                    },
                    "404": {
                        "description": "Table not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete table",
                "operationId": "e4e1cae6d3e81cd5d4a5096785af348a",
                "responses": {
                    "204": {
                        "description": "Table deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cat_name": {
                        "type": "string",
                        "example": "Coffee"
                    },
                    "cat_description": {
                        "type": "string",
                        "example": "Various coffee drinks"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "prod_name": {
                        "type": "string",
                        "example": "Espresso"
                    },
                    "prod_price": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "prod_description": {
                        "type": "string",
                        "example": "Rich and bold espresso"
                    },
                    "prod_image_url": {
                        "type": "string",
                        "example": "https://example.com/espresso.jpg"
                    },
                    "prod_is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "prod_meta": {
                        "type": "object"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}